+++ crates/snforge-scarb-plugin/src/attributes/available_gas.rs
@@ -1,12 +1,11 @@
 use crate::{
     args::Arguments,
     attributes::{AttributeCollector, AttributeInfo, AttributeTypeData},
-    branch,
     cairo_expression::CairoExpression,
     config_statement::extend_with_config_cheatcodes,
     types::{Number, ParseFromExpr},
 };
-use cairo_lang_macro::{Diagnostic, Diagnostics, ProcMacroResult, Severity, TokenStream, quote};
+use cairo_lang_macro::{Diagnostic, Diagnostics, ProcMacroResult, TokenStream, quote};
 use cairo_lang_parser::utils::SimpleParserDatabase;
 
 pub struct AvailableGasCollector;
@@ -25,8 +24,7 @@ impl AttributeCollector for AvailableGasCollector {
         args: Arguments,
         _warns: &mut Vec<Diagnostic>,
     ) -> Result<TokenStream, Diagnostics> {
-        let expr = branch!(from_resource_bounds(db, &args), from_max_gas(db, &args))?;
-        Ok(expr)
+        Ok(from_resource_bounds(db, &args)?)
     }
 }
 
@@ -63,29 +61,12 @@ fn from_resource_bounds(
     let l2_gas_expr = l2_gas.as_cairo_expression();
 
     Ok(quote!(
-        snforge_std::_internals::config_types::AvailableGasConfig::MaxResourceBounds(
-             snforge_std::_internals::config_types::AvailableResourceBoundsConfig {
-                 l1_gas: #l1_gas_expr,
-                 l1_data_gas: #l1_data_gas_expr,
-                 l2_gas: #l2_gas_expr,
-             }
-         )
+        snforge_std::_internals::config_types::AvailableResourceBoundsConfig {
+            l1_gas: #l1_gas_expr,
+            l1_data_gas: #l1_data_gas_expr,
+            l2_gas: #l2_gas_expr,
+        }
     ))
-}
-
-fn from_max_gas(db: &SimpleParserDatabase, args: &Arguments) -> Result<TokenStream, Diagnostic> {
-    let &[arg] = args
-        .unnamed_only::<AvailableGasCollector>()?
-        .of_length::<1, AvailableGasCollector>()?;
-
-    let gas =
-        Number::parse_from_expr::<AvailableGasCollector>(db, arg.1, arg.0.to_string().as_str())?;
-
-    gas.validate_in_gas_range::<AvailableGasCollector>("max_gas")?;
-
-    let gas = gas.as_cairo_expression();
-
-    Ok(quote!(snforge_std::_internals::config_types::AvailableGasConfig::MaxGas(#gas)))
 }
 
 #[must_use]