+++ crates/snforge-scarb-plugin/tests/integration/multiple_attributes.rs
@@ -112,13 +112,11 @@ fn works_with_few_attributes() {
                 if snforge_std::_internals::is_config_run() {
                     let mut data = array![];
 
-                    snforge_std::_internals::config_types::AvailableGasConfig::MaxResourceBounds(
-                        snforge_std::_internals::config_types::AvailableResourceBoundsConfig {
-                            l1_gas: 0x1,
-                            l1_data_gas: 0x2,
-                            l2_gas: 0x3
-                        }
-                    )
+                    snforge_std::_internals::config_types::AvailableResourceBoundsConfig {
+                        l1_gas: 0x1,
+                        l1_data_gas: 0x2,
+                        l2_gas: 0x3
+                    }
                     .serialize(ref data);
 
                     starknet::testing::cheatcode::<'set_config_available_gas'>(data.span());
@@ -144,13 +142,11 @@ fn works_with_few_attributes() {
                 if snforge_std::_internals::is_config_run() {
                     let mut data = array![];
 
-                    snforge_std::_internals::config_types::AvailableGasConfig::MaxResourceBounds(
-                        snforge_std::_internals::config_types::AvailableResourceBoundsConfig {
-                            l1_gas: 0x1,
-                            l1_data_gas: 0x2,
-                            l2_gas: 0x3
-                        }
-                    )
+                    snforge_std::_internals::config_types::AvailableResourceBoundsConfig {
+                        l1_gas: 0x1,
+                        l1_data_gas: 0x2,
+                        l2_gas: 0x3
+                    }
                     .serialize(ref data);
 
                     starknet::testing::cheatcode::<'set_config_available_gas'>(data.span());
@@ -240,13 +236,11 @@ fn works_with_fuzzer_config_wrapper() {
                 if snforge_std::_internals::is_config_run() {
                     let mut data = array![];
 
-                    snforge_std::_internals::config_types::AvailableGasConfig::MaxResourceBounds(
-                        snforge_std::_internals::config_types::AvailableResourceBoundsConfig {
-                            l1_gas: 0xffffffffffffffff,
-                            l1_data_gas: 0xffffffffffffffff,
-                            l2_gas: 0x3e7
-                        }
-                    )
+                    snforge_std::_internals::config_types::AvailableResourceBoundsConfig {
+                        l1_gas: 0xffffffffffffffff,
+                        l1_data_gas: 0xffffffffffffffff,
+                        l2_gas: 0x3e7
+                    }
                     .serialize(ref data);
 
                     starknet::testing::cheatcode::<'set_config_available_gas'>(data.span());
@@ -288,13 +282,11 @@ fn works_with_fuzzer_config_wrapper() {
                 if snforge_std::_internals::is_config_run() {
                     let mut data = array![];
 
-                    snforge_std::_internals::config_types::AvailableGasConfig::MaxResourceBounds(
-                        snforge_std::_internals::config_types::AvailableResourceBoundsConfig {
-                            l1_gas: 0xffffffffffffffff,
-                            l1_data_gas: 0xffffffffffffffff,
-                            l2_gas: 0x3e7
-                        }
-                    )
+                    snforge_std::_internals::config_types::AvailableResourceBoundsConfig {
+                        l1_gas: 0xffffffffffffffff,
+                        l1_data_gas: 0xffffffffffffffff,
+                        l2_gas: 0x3e7
+                    }
                     .serialize(ref data);
 
                     starknet::testing::cheatcode::<'set_config_available_gas'>(data.span());
@@ -326,13 +318,11 @@ fn works_with_fuzzer_config_wrapper() {
                 if snforge_std::_internals::is_config_run() {
                     let mut data = array![];
 
-                    snforge_std::_internals::config_types::AvailableGasConfig::MaxResourceBounds(
-                        snforge_std::_internals::config_types::AvailableResourceBoundsConfig {
-                            l1_gas: 0xffffffffffffffff,
-                            l1_data_gas: 0xffffffffffffffff,
-                            l2_gas: 0x3e7
-                        }
-                    )
+                    snforge_std::_internals::config_types::AvailableResourceBoundsConfig {
+                        l1_gas: 0xffffffffffffffff,
+                        l1_data_gas: 0xffffffffffffffff,
+                        l2_gas: 0x3e7
+                    }
                     .serialize(ref data);
 
                     starknet::testing::cheatcode::<'set_config_available_gas'>(data.span());
@@ -368,13 +358,11 @@ fn works_with_fuzzer_config_wrapper() {
                 if snforge_std::_internals::is_config_run() {
                     let mut data = array![];
 
-                    snforge_std::_internals::config_types::AvailableGasConfig::MaxResourceBounds(
-                        snforge_std::_internals::config_types::AvailableResourceBoundsConfig {
-                            l1_gas: 0xffffffffffffffff,
-                            l1_data_gas: 0xffffffffffffffff,
-                            l2_gas: 0x3e7
-                        }
-                    )
+                    snforge_std::_internals::config_types::AvailableResourceBoundsConfig {
+                        l1_gas: 0xffffffffffffffff,
+                        l1_data_gas: 0xffffffffffffffff,
+                        l2_gas: 0x3e7
+                    }
                     .serialize(ref data);
 
                     starknet::testing::cheatcode::<'set_config_available_gas'>(data.span());